+++ R/rexport.R
@@ -1,0 +1,24 @@
+
+#' @importFrom wrapr %.>%
+#' @export
+wrapr::`%.>%`
+
+#' @importFrom wrapr let
+#' @export
+wrapr::let
+
+#' @importFrom wrapr qc
+#' @export
+wrapr::qc
+
+#' @importFrom wrapr build_frame
+#' @export
+wrapr::build_frame
+
+#' @importFrom wrapr draw_frame
+#' @export
+wrapr::draw_frame
+
+#' @importFrom wrapr qchar_frame
+#' @export
+wrapr::qchar_frame